+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 Default configuration is now using the official OpenAPI example.
 
+### Enhancements ###
+
+- Logs are now cleared on each PyxelRest update
+
 ## 0.51 (2017-02-23) ##
 
 ### Enhancements ###